Refactor ActiveRitualTracker to use superclass abstract functionality that other events from other contracts can be easily tracked in the same way.
