+++ nucypher/blockchain/eth/trackers/dkg.py
@@ -1,76 +1,16 @@
-import datetime
-import os
-import time
-from typing import Callable, List, Optional, Tuple
+from typing import Optional, Tuple
 
 import maya
 from prometheus_client import REGISTRY, Gauge
-from twisted.internet import threads
 from web3.datastructures import AttributeDict
 
 from nucypher.blockchain.eth.models import Coordinator
-from nucypher.policy.conditions.utils import camel_case_to_snake
+from nucypher.blockchain.eth.trackers.events import EventTracker
+from nucypher.blockchain.eth.utils import get_block_just_before
 from nucypher.utilities.cache import TTLCache
-from nucypher.utilities.events import EventScanner, JSONifiedState
-from nucypher.utilities.logging import Logger
-from nucypher.utilities.task import SimpleTask
 
 
-class EventActuator(EventScanner):
-    """Act on events that are found by the scanner."""
-
-    def __init__(
-        self,
-        hooks: List[Callable],
-        clear: bool = True,
-        *args,
-        **kwargs,
-    ):
-        self.log = Logger("EventActuator")
-        if clear and os.path.exists(JSONifiedState.STATE_FILENAME):
-            os.remove(JSONifiedState.STATE_FILENAME)
-        self.hooks = hooks
-        super().__init__(*args, **kwargs)
-
-    def process_event(
-        self, event: AttributeDict, get_block_when: Callable[[int], datetime.datetime]
-    ):
-        for hook in self.hooks:
-            try:
-                hook(event, get_block_when)
-            except Exception as e:
-                self.log.warn("Error during event hook: {}".format(e))
-                raise
-        super().process_event(event, get_block_when)
-
-
-class EventScannerTask(SimpleTask):
-    """Task that runs the event scanner in a looping call."""
-
-    INTERVAL = 120  # seconds
-
-    def __init__(self, scanner: Callable):
-        self.scanner = scanner
-        super().__init__()
-
-    def run(self) -> None:
-        self.scanner()
-
-    def handle_errors(self, *args, **kwargs) -> None:
-        self.log.warn(
-            "Error during ritual event scanning: {}".format(args[0].getTraceback())
-        )
-        if not self._task.running:
-            self.log.warn("Restarting event scanner task!")
-            self.start(now=False)  # take a breather
-
-
-class ActiveRitualTracker:
-
-    CHAIN_REORG_SCAN_WINDOW = 20
-    MAX_CHUNK_SIZE = 10000
-    MIN_CHUNK_SIZE = 60  # 60 blocks @ 2s per block on Polygon = 120s of blocks (somewhat related to interval)
-
+class ActiveRitualTracker(EventTracker):
     # how often to check/purge for expired cached values - 8hrs?
     _PARTICIPATION_STATES_PURGE_INTERVAL = 60 * 60 * 8
 
@@ -99,46 +39,27 @@ class ActiveRitualTracker:
         operator,
         persistent: bool = False,  # TODO: use persistent storage?
     ):
-        self.log = Logger("RitualTracker")
-
-        self.operator = operator
-        self.coordinator_agent = operator.coordinator_agent
-
-        # Restore/create persistent event scanner state
-        self.persistent = persistent
-        self.state = JSONifiedState(persistent=persistent)
-        self.state.restore()
-
-        # Map events to handlers
-        self.actions = {
-            self.contract.events.StartRitual: self.operator.perform_round_1,
-            self.contract.events.StartAggregationRound: self.operator.perform_round_2,
+        contract = operator.coordinator_agent.contract
+        actions = {
+            contract.events.StartRitual: operator.perform_round_1,
+            contract.events.StartAggregationRound: operator.perform_round_2,
         }
-
-        self.events = [
-            self.contract.events.StartRitual,
-            self.contract.events.StartAggregationRound,
-            self.contract.events.EndRitual,
+        events = [
+            contract.events.StartRitual,
+            contract.events.StartAggregationRound,
+            contract.events.EndRitual,
         ]
 
-        # TODO: Remove the default JSON-RPC retry middleware
-        # as it correctly cannot handle eth_getLogs block range throttle down.
-        # self.web3.middleware_onion.remove(http_retry_request_middleware)
-
-        self.scanner = EventActuator(
-            hooks=[self._handle_ritual_event],
-            web3=self.web3,
-            state=self.state,
-            contract=self.contract,
-            events=self.events,
-            # How many maximum blocks at the time we request from JSON-RPC,
-            # and we are unlikely to exceed the response size limit of the JSON-RPC server
-            max_chunk_scan_size=self.MAX_CHUNK_SIZE,
-            min_chunk_scan_size=self.MIN_CHUNK_SIZE,
-            chain_reorg_rescan_window=self.CHAIN_REORG_SCAN_WINDOW,
+        super().__init__(
+            operator=operator,
+            web3=operator.coordinator_agent.blockchain.w3,
+            contract=contract,
+            events=events,
+            actions=actions,
+            persistent=persistent,
         )
 
-        self.task = EventScannerTask(scanner=self.scan)
+        self.coordinator_agent = operator.coordinator_agent
 
         cache_ttl = (
             self.coordinator_agent.get_timeout()
@@ -151,74 +72,26 @@ class ActiveRitualTracker:
             seconds=self._PARTICIPATION_STATES_PURGE_INTERVAL
         )
 
-    @property
-    def provider(self):
-        return self.web3.provider
-
-    @property
-    def web3(self):
-        return self.coordinator_agent.blockchain.w3
-
-    @property
-    def contract(self):
-        return self.coordinator_agent.contract
-
     # TODO: should sample_window_size be additionally configurable/chain-dependent?
     def _get_first_scan_start_block_number(self, sample_window_size: int = 100) -> int:
         """
         Returns the block number to start scanning for events from.
         """
-        w3 = self.web3
         timeout = self.coordinator_agent.get_timeout()
-
-        latest_block = w3.eth.get_block("latest")
-        if latest_block.number == 0:
-            return 0
-
-        # get average block time
-        sample_block_number = latest_block.number - sample_window_size
-        if sample_block_number <= 0:
-            return 0
-        base_block = w3.eth.get_block(sample_block_number)
-        average_block_time = (
-            latest_block.timestamp - base_block.timestamp
-        ) / sample_window_size
-
-        number_of_blocks_in_the_past = int(timeout / average_block_time)
-
-        expected_start_block = w3.eth.get_block(
-            max(0, latest_block.number - number_of_blocks_in_the_past)
+        return get_block_just_before(
+            w3=self.web3, how_far_back=timeout, sample_window_size=sample_window_size
         )
-        target_timestamp = latest_block.timestamp - timeout
 
-        # Keep looking back until we find the last block before the target timestamp
-        while (
-            expected_start_block.number > 0
-            and expected_start_block.timestamp > target_timestamp
-        ):
-            expected_start_block = w3.eth.get_block(expected_start_block.number - 1)
-
-        # if non-zero block found - return the block before
-        return expected_start_block.number - 1 if expected_start_block.number > 0 else 0
-
-    def start(self) -> None:
-        """Start the event scanner task."""
-        self.task.start()
-
-    def stop(self) -> None:
-        """Stop the event scanner task."""
-        self.task.stop()
-
-    def _action_required(self, ritual_event: AttributeDict) -> bool:
+    def _action_required(self, event: AttributeDict) -> bool:
         """Check if an action is required for a given ritual event."""
         # establish participation state first
-        participation_state = self._get_participation_state(ritual_event)
+        participation_state = self._get_participation_state(event)
 
         if not participation_state.participating:
             return False
 
         # does event have an associated action
-        event_type = getattr(self.contract.events, ritual_event.event)
+        event_type = getattr(self.contract.events, event.event)
 
         event_has_associated_action = event_type in self.actions
         already_posted_transcript = (
@@ -374,82 +247,8 @@ class ActiveRitualTracker:
 
         return participation_state
 
-    def __execute_action(
-        self,
-        ritual_event: AttributeDict,
-        timestamp: int,
-        defer: bool = False,
-    ):
-        """Execute a round of a ritual asynchronously."""
-        # NOTE: this format splits on *capital letters* and converts to snake case
-        #  so "StartConfirmationRound" becomes "start_confirmation_round"
-        #  do not use abbreviations in event names (e.g. "DKG" -> "d_k_g")
-        formatted_kwargs = {
-            camel_case_to_snake(k): v for k, v in ritual_event.args.items()
-        }
-        event_type = getattr(self.contract.events, ritual_event.event)
-
-        def task():
-            self.actions[event_type](timestamp=timestamp, **formatted_kwargs)
-
-        if defer:
-            d = threads.deferToThread(task)
-            d.addErrback(self.task.handle_errors)
-            return d
-        else:
-            return task()
-
-    def _handle_ritual_event(
-        self,
-        ritual_event: AttributeDict,
-        get_block_when: Callable[[int], datetime.datetime],
-    ):
-        # is event actionable
-        if not self._action_required(ritual_event):
-            self.log.debug(
-                f"Event '{ritual_event.event}', does not require further action either: not participating in ritual, no corresponding action or previously handled; skipping"
-            )
-            return
-
-        timestamp = int(get_block_when(ritual_event.blockNumber).timestamp())
-        d = self.__execute_action(ritual_event=ritual_event, timestamp=timestamp)
-        return d
-
-    def __scan(self, start_block, end_block, account):
-        # Run the scan
-        self.log.debug(
-            f"({account[:8]}) Scanning events in block range {start_block} - {end_block}"
-        )
-        start = time.time()
-        result, total_chunks_scanned = self.scanner.scan(start_block, end_block)
-        if self.persistent:
-            self.state.save()
-        duration = time.time() - start
-        self.log.debug(
-            f"Scanned total of {len(result)} events, in {duration} seconds, "
-            f"total {total_chunks_scanned} chunk scans performed"
-        )
-
     def scan(self):
-        """
-        Assume we might have scanned the blocks all the way to the last Ethereum block
-        that mined a few seconds before the previous scan run ended.
-        Because there might have been a minor Ethereum chain reorganisations since the last scan ended,
-        we need to discard the last few blocks from the previous scan results.
-        """
         last_scanned_block = self.scanner.get_last_scanned_block()
         self._LAST_SCANNED_BLOCK_METRIC.set(last_scanned_block)
 
-        if last_scanned_block == 0:
-            # first run so calculate starting block number based on dkg timeout
-            suggested_start_block = self._get_first_scan_start_block_number()
-        else:
-            self.scanner.delete_potentially_forked_block_data(
-                last_scanned_block - self.scanner.chain_reorg_rescan_window
-            )
-            suggested_start_block = self.scanner.get_suggested_scan_start_block()
-
-        end_block = self.scanner.get_suggested_scan_end_block()
-        self.__scan(
-            suggested_start_block, end_block, self.operator.transacting_power.account
-        )
+        super().scan()
